backend/requirements.txt
# Django core
Django==5.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
gunicorn==21.2.0

# CORS & Security
django-cors-headers==4.3.1

# Telegram
aiogram==3.3.0

# Google Sheets
google-api-python-client==2.114.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# Utilities
httpx==0.26.0
